add custom error messages for wrong login/password and token expired
