+++ app.js
@@ -11,7 +11,6 @@ const logger = require("./lib/logger");
 const login = require("./routes/login");
 const courses = require("./routes/courses");
 const media = require("./routes/media");
-const feed = require("./routes/feed");
 
 const app = express();
 
@@ -26,7 +25,6 @@ app.use("/api/v1/login", login);
 app.use(auth);
 app.use("/api/v1/courses", courses);
 app.use("/api/v1/media", media);
-app.use("/api/v1/feed", feed);
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
