+++ lib/auth.js
@@ -1,6 +1,7 @@
 "use strict";
 
 var njwt = require("njwt");
+const ERRORS = require("../lib/errors");
 
 var SIGNING_KEY = process.env.SIGNING_KEY;
 
@@ -18,7 +19,10 @@ module.exports = (req, res, next) => {
     njwt.verify(token, SIGNING_KEY, (err, ver) => {
       if (err) {
         // token is expired
-        sendError("Token is expired.", 401, next);
+        const message = ERRORS.TOKEN_EXPIRED;
+        const code = 401;
+        res.writeHead(code, message, {"content-type": "text/plain"});
+        res.end(message); // need to end the conection here, otherwise the error message is switched
       } else {
         // token is Gucci!
         req.user = ver.body;
@@ -27,6 +31,6 @@ module.exports = (req, res, next) => {
     });
   } else {
     // token not sent
-    sendError("Token was not received. Expected token in Authorization header.", 401, next);
+    sendError(ERRORS.TOKEN_NOT_RECEIVED, 401, next);
   }
 };
