+++ lib/mediaApi.js
@@ -40,7 +40,7 @@ function getLectures(courseName, callback) {
   });
   logger.info(`GET/ ${options.path}`);
   const request = http.get(options, res => processData(res, data => callback(null, data)));
-  request.setTimeout(900, () => {
+  request.setTimeout(500, () => {
     callback(`Request to ${options.hostname}/${options.path} timed out`);
   });
 }
@@ -58,7 +58,7 @@ function getLectureMetaData(semester, courseName, lectureName, callback) {
   });
 
   const require = http.get(options, res => processData(res, data => callback(null, data)));
-  require.setTimeout(500, () => {
+  require.setTimeout(750, () => {
     callback(`Request to ${options.hostname}/${options.path} timed out`);
   });
 }
