+++ package.json
@@ -12,6 +12,7 @@
     "cookie-parser": "~1.3.5",
     "debug": "~2.2.0",
     "express": "~4.13.1",
+    "keymirror": "^0.1.1",
     "mysql": "^2.10.2",
     "njwt": "^0.2.3",
     "winston": "^2.2.0",
