+++ sql/course-list-meta-data.sql
