+++ sql/id-and-hash-from-email.sql
